+++ src/Class/Service/BulkImport/BulkImportExecutorInterface.php
@@ -1,0 +1,40 @@
+<?php
+declare(strict_types=1);
+
+/**
+ * This source file is available under the terms of the
+ * Pimcore Open Core License (POCL)
+ * Full copyright and license information is available in
+ * LICENSE.md which is distributed with this source code.
+ *
+ *  @copyright  Copyright (c) Pimcore GmbH (https://www.pimcore.com)
+ *  @license    Pimcore Open Core License (POCL)
+ */
+
+namespace Pimcore\Bundle\StudioBackendBundle\Class\Service\BulkImport;
+
+use Pimcore\Bundle\StudioBackendBundle\Class\Util\ClassDefinitionType;
+use Pimcore\Bundle\StudioBackendBundle\Exception\Api\ElementSavingFailedException;
+use Pimcore\Bundle\StudioBackendBundle\Exception\Api\EnvironmentException;
+use Pimcore\Bundle\StudioBackendBundle\Exception\Api\InvalidArgumentException;
+use Pimcore\Bundle\StudioBackendBundle\Exception\Api\NotFoundException;
+use Pimcore\Model\UserInterface;
+
+/**
+ * @internal
+ */
+interface BulkImportExecutorInterface
+{
+    /**
+     * @throws ElementSavingFailedException
+     * @throws EnvironmentException
+     * @throws InvalidArgumentException
+     * @throws NotFoundException
+     */
+    public function importSingleItem(
+        ClassDefinitionType $type,
+        string $name,
+        array $exportEntry,
+        UserInterface $user,
+    ): void;
+}
